fix compile on 32-bit linux (issue with 32-bit fields in libc statfs)

+++ example/src/passthrough.rs
@@ -81,11 +81,11 @@ fn statfs_to_fuse(statfs: libc::statfs) -> Statfs {
 #[cfg(target_os = "linux")]
 fn statfs_to_fuse(statfs: libc::statfs) -> Statfs {
     Statfs {
-        blocks: statfs.f_blocks,
-        bfree: statfs.f_bfree,
-        bavail: statfs.f_bavail,
-        files: statfs.f_files,
-        ffree: statfs.f_ffree,
+        blocks: statfs.f_blocks as u64,
+        bfree: statfs.f_bfree as u64,
+        bavail: statfs.f_bavail as u64,
+        files: statfs.f_files as u64,
+        ffree: statfs.f_ffree as u64,
         bsize: statfs.f_bsize as u32,
         namelen: statfs.f_namelen as u32,
         frsize: statfs.f_frsize as u32,
